refactor(tx-manager): share runtime timeout helper
